requirements.txt
asgiref
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
drf-yasg
celery
redis
requests
Pillow
sqlparse
typing_extensions
pytz
sqlalchemy
pydantic
python-dotenv